Add VodTvSectionFragment
Add a base class in which users might be able to playback VOD content and
save it's content so it can used as a cache.

ContentPersistence: Change the return type of size() from int to long since
there might be a lot of values (I doubt there will be more than the max
value of an Integer but it is more as a requirement from some databases)

+++ library/src/main/java/com/zaclimon/iptvproviderlibrary/persistence/ContentPersistence.java
@@ -78,7 +78,7 @@ public interface ContentPersistence {
      *
      * @return the total number of elements for this medium.
      */
-    int size();
+    long size();
 
     /**
      * Gets the size for a given category in the persistence medium
@@ -86,5 +86,5 @@ public interface ContentPersistence {
      * @param category the category in which a size is wanted
      * @return the total number of elements for that given category.
      */
-    int size(String category);
+    long size(String category);
 }
